requirements.txt
eca_lib==0.1.6
sentence_transformers==5.0.0
SQLAlchemy==2.0.43
